+++ regimes/fr/fr.go
@@ -41,9 +41,6 @@ func New() *tax.RegimeDef {
 			`),
 		},
 		TimeZone: "Europe/Paris",
-		Scenarios: []*tax.ScenarioSet{
-			invoiceScenarios,
-		},
 		Corrections: []*tax.CorrectionDefinition{
 			{
 				Schema: bill.ShortSchemaInvoice,